+++ wgpu-hal/src/vulkan/device.rs
@@ -1103,8 +1103,6 @@ impl crate::Device<super::Api> for super::Device {
         &self,
         desc: &crate::SamplerDescriptor,
     ) -> Result<super::Sampler, crate::DeviceError> {
-        let lod_range = desc.lod_clamp.clone().unwrap_or(0.0..16.0);
-
         let mut vk_info = vk::SamplerCreateInfo::builder()
             .flags(vk::SamplerCreateFlags::empty())
             .mag_filter(conv::map_filter_mode(desc.mag_filter))
@@ -1113,8 +1111,8 @@ impl crate::Device<super::Api> for super::Device {
             .address_mode_u(conv::map_address_mode(desc.address_modes[0]))
             .address_mode_v(conv::map_address_mode(desc.address_modes[1]))
             .address_mode_w(conv::map_address_mode(desc.address_modes[2]))
-            .min_lod(lod_range.start)
-            .max_lod(lod_range.end);
+            .min_lod(desc.lod_clamp.start)
+            .max_lod(desc.lod_clamp.end);
 
         if let Some(fun) = desc.compare {
             vk_info = vk_info
@@ -1122,16 +1120,14 @@ impl crate::Device<super::Api> for super::Device {
                 .compare_op(conv::map_comparison(fun));
         }
 
-        if let Some(aniso) = desc.anisotropy_clamp {
-            if self
-                .shared
-                .downlevel_flags
-                .contains(wgt::DownlevelFlags::ANISOTROPIC_FILTERING)
-            {
-                vk_info = vk_info
-                    .anisotropy_enable(true)
-                    .max_anisotropy(aniso.get() as f32);
-            }
+        if self
+            .shared
+            .downlevel_flags
+            .contains(wgt::DownlevelFlags::ANISOTROPIC_FILTERING)
+        {
+            vk_info = vk_info
+                .anisotropy_enable(true)
+                .max_anisotropy(desc.anisotropy_clamp);
         }
 
         if let Some(color) = desc.border_color {